pyproject.toml
[project]
name = "max-benchmark"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Ryosuke Kamesawa", email = "emak.ryo@gmail.com" }
]
dependencies = [
    "max-engine>=24.4.0",
    "torch==2.2.2+cpu",
    "transformers==4.40.1",
    "numpy<2.0",
    "torchvision==0.17.2+cpu",
    "onnxruntime>=1.18.1",
    "onnx>=1.16.1",
    "pandas>=2.2.2",
    "timm>=1.0.7",
    "jupyterlab>=4.0.13",
    "matplotlib>=3.9.1",
    "ultralytics==8.2.5",
]
readme = "README.md"
requires-python = ">= 3.10"

[project.scripts]
benchmark = "max_benchmark:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[[tool.rye.sources]]
name = "modular"
# Replace MAX wheel path with your own.
# Run `modular config max.path` to get the path.
# Typically, it is `~/.modular/pkg/packages.modular.com_max/wheels`.
url = "file:///<HOME>/.modular/pkg/packages.modular.com_max/wheels"
type = "find-links"

[[tool.rye.sources]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/max_benchmark"]
